Split Operators file into Operators, Classification, image and point
