+++ source/vision/algorithms/gpu/shaders/blur_fs.glsl
