+++ source/vision/algorithms/gpu/shaders/passthrough_fs.glsl
