+++ source/vision/algorithms/gpu/shaders/passthrough_vs.glsl
