+++ source/vision/algorithms/gpu/shaders/segment_fs.glsl
