+++ source/vision/algorithms/gpu/shaders/vision_gpu_shaders.qrc
