+++ source/vision/algorithms/original/algorithm_original.hpp
@@ -6,9 +6,10 @@
 #include <QMutex>
 #include <QVideoFrame>
 
+#include "../shared/operators.hpp"
+#include "../vision_algorithm.hpp"
 #include "movement3d/movement3d.hpp"
 #include "movement3d/movement3d_filter.hpp"
-#include "vision_algorithm.hpp"
 
 class AlgorithmOriginal : public VisionAlgorithm {
     public:
