+++ source/vision/vision.hpp
@@ -15,9 +15,9 @@
 #include <QVideoFrame>
 
 #include "acquisition.hpp"
-#include "algorithms/algorithm_gpu/algorithm_gpu.hpp"
-#include "algorithms/algorithm_original.hpp"
-#include "algorithms/algorithm_random.hpp"
+#include "algorithms/gpu/algorithm_gpu.hpp"
+#include "algorithms/original/algorithm_original.hpp"
+#include "algorithms/random/algorithm_random.hpp"
 #include "algorithms/vision_algorithm.hpp"
 #include "augmentation_widget/augmentation_widget.hpp"
 #include "movement3d/movement3d.hpp"
